requirements.txt
fastapi 
uvicorn[standard] 
 
# Pydantic 
pydantic 
pydantic-settings 
 
# LangChain + OpenAI 
langchain 
langchain-openai 
langchain-community 
 
# Vector store 
chromadb 
 
# Database 
sqlalchemy 
psycopg2-binary 
 
# Env 
python-dotenv 
 
# PDF / docs 
pymupdf 
unstructured
